fix WikiLinkExtractor class
